refactor test
refactored interface. updated packages and Dockerfile

updated code
