+++ .pre-commit-config.yaml
@@ -3,9 +3,9 @@
 
 
 default_language_version:
-    python: python3.10.2
+    python: python3.11.3
 default_stages: [commit, commit-msg, merge-commit, push]
-
+fail_fast: true
 repos:
 -   repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v4.4.0
@@ -24,7 +24,7 @@ repos:
         types: [file, python]
         args: [
             --line-length=79,
-            --target-version=py310
+            --target-version=py311
         ]
 -   repo: https://github.com/pre-commit/mirrors-mypy
     rev: v1.3.0
