+++ Dockerfile
@@ -1,5 +1,5 @@
 # Starts core package
-FROM python:3.10.2
+FROM python:3.11.3
 ENV PYTHONDONTWRITEBYTECODE 1
 ENV PYTHONUNBUFFERED 1
