+++ docker-compose.yaml
@@ -26,7 +26,7 @@ services:
     restart: always
     container_name: rmq
     build:
-      context: .
+      # context: .
       dockerfile: Dockerfile
     hostname: rmq
     links:
