+++ test/conftest.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import logging
 import random
 import uuid
@@ -15,8 +17,8 @@ from typing import (
 import pytest
 from pytest import FixtureRequest
 
-from core.interface import ITaskQueueRepository
-from core.taskman.domain.entity import TaskQueue, Task, TaskWorker
+from core.taskman.interface import ITaskQueueRepository
+from core.taskman.domain.entity import TaskQueue, Task, TaskWorker, TaskStatus
 from core.taskman.domain.services import QueuingService
 
 pytest_logger = logging.getLogger(__file__)
@@ -57,15 +59,18 @@ stub_num_workers = 2
 
 
 def universal_uuid() -> uuid.UUID:
+    """Generate a uuid5 wiht a deterministic seed."""
     random.seed(0)
     return uuid.uuid5(uuid.NAMESPACE_DNS, str(random.random().hex))
 
 
 def func_descriptor() -> Callable[[List[int]], int]:
+    """Return a function with a fixed descriptor."""
     return lambda x_list: sum(x_list)
 
 
 def stub_args() -> Tuple[int, ...]:
+    """Returns stubbed arguments."""
     return 1, 2, 3, 4, 5
 
 
@@ -76,6 +81,7 @@ def stub_task_param_dict() -> (
         pytest.fixture,
     ]
 ):
+    """Returns a dictionary of stubbed attribtues."""
     return dict(
         task_id=universal_uuid(),
         task_func=func_descriptor,
@@ -98,7 +104,36 @@ class DummyTaskQueueRepo(ITaskQueueRepository):
     ) -> None:
         """Create a queue."""
         pytest_logger.info("creating queue: %s", queue_name)
-        pass
+        task_queue = TaskQueue(
+            queue_name=queue_name,
+            num_task_workers=stub_num_workers,
+        )
+        cls.QueueMockDb[queue_name] = task_queue
+
+    @classmethod
+    def add_task(
+        cls,
+        queue_name: str,
+        func_descriptor: Callable[..., Any],
+        /,
+        *task_args: Tuple[Any, ...],
+        task_status: TaskStatus = TaskStatus.ENQUEUED,
+        **task_opts: str
+        | Sequence[str]
+        | Sequence[TaskWorker]
+        | Sequence[Task],
+    ) -> None:
+        if not (task_queue := cls.get(queue_name)):
+            raise ValueError(f"Queue {queue_name} does not exist.")
+        task_queue.tasks += [
+            Task(
+                task_id=universal_uuid(),
+                task_func=func_descriptor,
+                args=task_args,
+                status=task_status,
+                options=task_opts.get("options", {}),
+            )
+        ]
 
     @classmethod
     def get(
@@ -109,22 +144,7 @@ class DummyTaskQueueRepo(ITaskQueueRepository):
     ) -> TaskQueue | None:
         """Get a queue."""
         pytest_logger.info("getting queue: %s", queue_name)
-
-        if queue_name in cls.QueueMockDb:
-            return cls.QueueMockDb[queue_name]
-        task = Task(
-            task_id=universal_uuid(),
-            task_func=func_descriptor,
-            args=stub_args(),
-            **options,
-        )
-        task_queue = TaskQueue(
-            queue_name=queue_name,
-            num_task_workers=stub_num_workers,
-            tasks=[task],
-        )
-        cls.QueueMockDb[queue_name] = task_queue
-        return task_queue
+        return cls.QueueMockDb.get(queue_name)
 
     @classmethod
     def clear(cls) -> None:
